% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{gainFactors}
\alias{gainFactors}
\title{Return GAMMA}
\usage{
gainFactors(Number, BURNIN)
}
\arguments{
\item{Number}{Number}

\item{BURNIN}{BURNIN.}
}
\value{
GAMMA, a vector of gain factors
}
\description{
Generate a series of gain factors.
}
