% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_Code.R
\name{generateSimData}
\alias{generateSimData}
\title{Data for simulations.}
\usage{
generateSimData(Groups = 5, Dimensions = 5, Number = 10^4)
}
\arguments{
\item{Groups}{Number of mixture components. Default 5.}

\item{Dimensions}{number of Dimensions. Default 5.}

\item{Number}{number of samples. Default 10^4.}
}
\value{
List of results: X, Y, simobject.
}
\description{
Wrapper for MixSim
}
\examples{
sims<-generateSimData(Groups=10, Dimensions=10, Number=10^4)
sims<-generateSimData()

}
