% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mahalanobis_HD}
\alias{mahalanobis_HD}
\title{Mahalanobis Distance}
\usage{
mahalanobis_HD(y, mu, sigma)
}
\arguments{
\item{y}{vector of length n.}

\item{mu}{vector of length n.}

\item{sigma}{nxn covariance matrix.}
}
\value{
Distance (scalar)
}
\description{
Calculates Mahalanobis distance between y and mu, scaled by sigma.
}
