% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{norm_HD}
\alias{norm_HD}
\title{Euclidean Distance}
\usage{
norm_HD(y, mu, sigma)
}
\arguments{
\item{y}{vector of length n.}

\item{mu}{vector of length n.}

\item{sigma}{nxn covariance matrix.}
}
\value{
distance (scalar).
}
\description{
Calculates Euclidean distance between y and mu, scaled by sigma.
}
