% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{package}
\name{SAGMM}
\alias{SAGMM}
\alias{SAGMM-package}
\title{SAGMM: A package for Clustering via Stochastic Approximation and Gaussian Mixture Models.}
\description{
The SAGMM package allows for computation
}
\references{
Nguyen, H. D. & Jones, A. T. (2018). Big Data-appropriate clustering via stochastic approximation and Gaussian mixture models. In M. Ahmed & A.-S. K. Pathan (Eds.), Data Analytics:Concepts, Techniques, and Applications. Boca Raton: CRC Press.
}
\author{
Andrew T. Jones and Hien D. Nguyen
}
