% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{SAGMMFit}
\alias{SAGMMFit}
\title{Clustering via Stochastic Approximation and Gaussian Mixture Models}
\usage{
SAGMMFit(X, Y, MS, BURNIN = 5, Groups = 5, kstart = 10, plot = F)
}
\arguments{
\item{X}{numeric maxtrix of the data.}

\item{Y}{True classes.}

\item{MS}{MixSim Simulation object.}

\item{BURNIN}{Ratio of observations to use as a burn in before algorithm begins.}

\item{Groups}{Number of mixture components.}

\item{kstart}{number of kmeans starts to initialise.}

\item{plot}{If TRUE generates a plot of the clustering.}
}
\value{
List of results
}
\description{
Fit a GMM with SA.
}
\examples{
sims<-generateSimData(Groups=10, Dimensions=10, Number=10^4)
res1<-SAGMMFit(sims$X, sims$Y, sims$MS, mode="C")

}
