% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{gainFactors}
\alias{gainFactors}
\title{Return Gamma, a sequence of gain factors}
\usage{
gainFactors(Number, Burnin)
}
\arguments{
\item{Number}{Number of values required.}

\item{Burnin}{Number of 'Burnin' values at the beginning of sequence.}
}
\value{
Gamma, a vector of gain factors.
}
\description{
Generate a series of gain factors.
}
\examples{
g<-gainFactors(10^4, 2*10^3)
}
