% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{gainFactors}
\alias{gainFactors}
\title{Return GAMMA}
\usage{
gainFactors(Number, BURNIN)
}
\arguments{
\item{Number}{Number of values required.}

\item{BURNIN}{Number of 'burnin' values at the beginning of sequence}
}
\value{
GAMMA, a vector of gain factors
}
\description{
Generate a series of gain factors.
}
\examples{
g<-gainFactors(10^4, 2*10^3)
}
