% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{generateSimData}
\alias{generateSimData}
\title{Generate data for simulations to test the SAGMM package..}
\usage{
generateSimData(ngroups = 5, Dimensions = 5, Number = 10^4)
}
\arguments{
\item{ngroups}{Number of mixture components. Default 5.}

\item{Dimensions}{number of Dimensions. Default 5.}

\item{Number}{number of samples. Default 10^4.}
}
\value{
List of results: X, Y, simobject.
}
\description{
This function is primarily a convienence wrapper for MixSim.
}
\examples{
sims<-generateSimData(ngroups=10, Dimensions=10, Number=10^4)
sims<-generateSimData()
}
