% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{main_loop}
\alias{main_loop}
\title{Main Computation Loop}
\usage{
main_loop(Number, Groups, PISTAR_O, MU_O, LAMBDA_O, GAMMA, X, Dimensions,
  SIGMA)
}
\arguments{
\item{Number}{Number of observations.}

\item{Groups}{Number of groups.}

\item{MU_O}{Matrix of group means}

\item{LAMBDA_O}{vector of lambda values. See Reference.}

\item{GAMMA}{vector of gamma values.See Reference.}

\item{X}{Matrix of Observations}

\item{Dimensions}{Dimension of observations.}

\item{PISTAR}{cube of pi* values. See Reference.}
}
\value{
Results List
PI - Group proportions.
MU - Group means. 
LAMBDA - Updated Lambda values.
LogLike - Logliklihood value.
TauMat - Conditional probabilities.
}
\description{
Performs primary compuation process. See Reference.
}
\references{
Nguyen & Jones (2018). Big Data-Appropriate Clustering via Stochastic Approximation and Gaussian Mixture Models. In Data Analytics (pp. 79-96). CRC Press.
}
\author{
Andrew T. Jones and Hien D. Nguyen
}
