% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{main_loop}
\alias{main_loop}
\title{Main Computation Loop}
\usage{
main_loop(Number, Groups, PISTAR_O, MU_O, LAMBDA_O, GAMMA, X, Dimensions,
  SIGMA)
}
\value{
RETURNVAL
}
\description{
THIS FUNCTION DOES
}
