% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{main_loop}
\alias{main_loop}
\title{Main Computation Loop}
\usage{
main_loop(Number, Groups, PISTAR_O, MU_O, LAMBDA_O, GAMMA, X, Dimensions,
  SIGMA)
}
\arguments{
\item{Number}{int ???}

\item{Groups}{int}

\item{PISTAR_O}{vector}

\item{MU_O}{matrix}

\item{LAMBDA_O}{vector}

\item{GAMMA}{vector}

\item{X}{matrix}

\item{Dimensions}{int}

\item{SIGMA}{nxnx? cube}
}
\value{
Results List
PI -
MU -
LAMBDA -
Comps - 
LogLike - 
TauMat -
}
\description{
Performs primary compuation process
}
