% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{package}
\name{SAGMM}
\alias{SAGMM}
\alias{SAGMM-package}
\title{SAGMM: A package for Clustering via Stochastic Approximation and Gaussian Mixture Models.}
\description{
The SAGMM package allows for computation of gaussian mixture models using stochastic approximation to increase efficiency with large data sets.
The primary function \code{SAGMMFit} allows this to be performed in a relative flexible manner.
}
\references{
Nguyen & Jones (2018). Big Data-Appropriate Clustering via Stochastic Approximation and Gaussian Mixture Models. In Data Analytics (pp. 79-96). CRC Press.
}
\author{
Andrew T. Jones and Hien D. Nguyen
}
