% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{SAGMMFit}
\alias{SAGMMFit}
\title{Clustering via Stochastic Approximation and Gaussian Mixture Models}
\usage{
SAGMMFit(X, Y, BURNIN = 5, Groups = 5, kstart = 10, plot = F)
}
\arguments{
\item{X}{numeric maxtrix of the data.}

\item{BURNIN}{Ratio of observations to use as a burn in before algorithm begins.}

\item{Groups}{Number of mixture components.}

\item{kstart}{number of kmeans starts to initialise.}

\item{plot}{If TRUE generates a plot of the clustering.}
}
\value{
A list containing
\item{Cluster}{The clustering of each observation.}
\item{plot}{A plot of the clustering (if requested).}
\item{l2}{Lambda^2}
\item{ARI1}{}
\item{ARI2}{}
\item{KM}{K-means clustering of the data.}
\item{ARI3}{}
\item{pi}{The cluster proportions}
\item{tau}{Tau matrix.}
\item{fit}{Output from C++ loop.}
}
\description{
Fit a GMM with SA.
}
\examples{
sims<-generateSimData(Groups=10, Dimensions=10, Number=10^4)
res1<-SAGMMFit(sims$X, sims$Y)

}
